+++ RELEASE_NOTES.md
@@ -1,8 +1,8 @@
 ### New Features
 
- - Implement V2 Get Order and Get all Orders
+- Implement V2 Order Products API
+- Implement V2 Order Shipping Addresses API (read-only)
 
-### Fixes
 
- - Fixed error where amount was `int` instead of `float` in price rules
+